Bl fixed

+++ BL/BusinessLogic/BinBusinessLogic.cs
@@ -22,30 +22,27 @@ namespace BL
         }
         //TODO - Implement Methods of GET/SET/UPDATE/DELETE things from the database
 
-        // TODO - change method's name to GetAllBins
-        // TODO - Some sort of inner join with Building to get adresses 
         public List<BinData> GetAllBinsData()
         {
-            List<Bin> dbBins = null;
+            List<spBin_GetBinListFullDetails_Result> dbBins = null;
             List<BinData> bins = new List<BinData>();
 
             try
             {
-                dbBins = this.db.Bins.ToList();
+                dbBins = this.db.spBin_GetBinListFullDetails().ToList();
 
-                foreach (Bin dbBin in dbBins)
+                foreach (spBin_GetBinListFullDetails_Result dbBin in dbBins)
                 {
                     BinData binData = new BinData()
                     {
                         binId = dbBin.BinId,
                         binTypeId = dbBin.BinTypeId,
-                        binTypeDesc = dbBin.LUT_BinType.BinTypeDesc,
-                        //cityAddress = dbBin.CityAddress,
+                        binTypeDesc = dbBin.BinTypeDesc,
+                        cityAddress = dbBin.AreaDesc,
                         currentCapacity = dbBin.CurrentCapacity,
-                        isInUser = dbBin.IsInUse,
-                        maxCapacity = dbBin.LUT_BinType.Capacity,
-                        //streetAddress = dbBin.StreetAddress,
-                        //streetNumber = dbBin.StreetNumber
+                        maxCapacity = dbBin.Capacity,
+                        streetAddress = dbBin.BuildingAddress,
+                        binTrashDisposalArea = dbBin.BinTrashDisposalArea,
                     };
 
                     bins.Add(binData);
@@ -58,8 +55,7 @@ namespace BL
             }
         }
 
-        // TODO - consider removing this method
-        //          Keep it only if it has an inner use.
+
         public List<Bin> GetAllBins()
         {
             try
@@ -72,11 +68,11 @@ namespace BL
             }
         }
 
-        public Bin GetBin(int id)
+        public Bin GetBin(int binId)
         {
             try
             {
-                return this.db.Bins.Where(x => x.BinId == id).SingleOrDefault();
+                return this.db.Bins.Where(x => x.BinId == binId).SingleOrDefault();
             }
             catch (Exception ex)
             {
@@ -89,7 +85,7 @@ namespace BL
             try
             {
                 this.db.Bins.Add(newBin);
-                this.db.SaveChanges();      //TODO - check if newBin got an id from db 
+                this.db.SaveChanges();      
             }
             catch (Exception ex)
             {
@@ -99,25 +95,19 @@ namespace BL
             return newBin;
         }
 
-        // TODO - the input parameter will be BinData so the method needs to be changed accordingly.
-        // TODO - line "oldBin = updatedBin;" is not good. 
-        //        Instead, assign each field seperately and only fields that need to be updated.
-        //        Especially not 'navigation methods'
-        public void UpdateBin(Bin updatedBin, DateTime dt)
+        public void UpdateBin(BinData updatedBin, DateTime dt)
         {
             try
-            {
-                Bin oldBin = this.db.Bins.Where(x => x.BinId == updatedBin.BinId).SingleOrDefault();
-                if (oldBin != null)
+            {   
+                Bin oldBin = this.db.Bins.Where(x => x.BinId == updatedBin.binId).SingleOrDefault();
+                if (oldBin == null)
                 {
-                    //oldBin.CurrentCapacity = updatedBin.CurrentCapacity;
-
-
-                    oldBin = updatedBin; // <- NOT GOOD - SEE COMMENT ABOVE !!!
-                    
-                    //you can update more properties here.... (in the same way)
-                    UpdateBinLogForEachBinAction(updatedBin, dt);
+                    throw new Exception("Bin not found.");
                 }
+
+                oldBin = BinDataToDbBin(updatedBin);
+
+                AddNewBinLog(oldBin, dt);         
                 this.db.SaveChanges();
             }
             catch (Exception ex)
@@ -126,10 +116,8 @@ namespace BL
             }
         }
 
-        // TODO - Change Method's name to AddNewBinLog
-        private void UpdateBinLogForEachBinAction(Bin updatedBin, DateTime dt)
+        private void AddNewBinLog(Bin updatedBin, DateTime dt)
         {
-            // TODO: Think about moving the code below into a class ... :S
 
             BinLog bLog = new BinLog();
 
@@ -148,14 +136,15 @@ namespace BL
             }
         }
 
-        // TODO - change input parameter name from 'id' to 'binId'
-        // TODO - when getting 'binToRemove' you used SingleOrDefault(); - that means result could be null/
-        //        you need to add validation that if binToRemove is null throw an exception - 'not found...'
-        public void DeleteBin(int id)
+        public void DeleteBin(int binId)
         {
             try
             {
-                Bin binToRemove = this.db.Bins.Where(x => x.BinId == id).SingleOrDefault();
+                Bin binToRemove = this.db.Bins.Where(x => x.BinId == binId).SingleOrDefault();
+                if (binToRemove == null)
+                {
+                    throw new Exception("Bin not found");
+                }
                 this.db.Bins.Remove(binToRemove);
                 this.db.SaveChanges();
             }
@@ -165,20 +154,6 @@ namespace BL
             }
         }
 
-        // TODO - move the validation to inside the try.
-        //        follow this pseudocode:
-        //
-        //          try
-        //          {
-        //              get bt.
-        //              check if bt you got is null
-        //                  if is null - throw an exception
-        //              return bt.capacity
-        //          }
-        //          catch 
-        //          { 
-        //              Errorhandler.handle.....
-        //          }
         public double GetMaxCapacityByBinType(int binTypeId)
         {
             LUT_BinType bt;
@@ -186,18 +161,48 @@ namespace BL
             try
             {
                 bt = this.db.LUT_BinType.Where(x => x.BinTypeId == binTypeId).SingleOrDefault();
+                if (bt == null)
+                {
+                    throw new Exception("There isn't id like this in the system.");
+                }
             }
             catch (Exception ex)
             {
                 throw ErrorHandler.Handle(ex, this);
             }
 
-            if (bt == null)
-            {
-                throw ErrorHandler.Handle(new Exception("There isn't id like this in the system."), this);
-            }
-
             return bt.Capacity;
+        }
+
+        public BinData DbBinToBinData(Bin dbBin)
+        {
+            List<spBin_GetBinListFullDetails_Result> dbBins = this.db.spBin_GetBinListFullDetails().ToList();
+            spBin_GetBinListFullDetails_Result currentBin = dbBins.Find(x => x.BinId == dbBin.BinId);
+            BinData binData = new BinData()
+            {
+                binId = currentBin.BinId,
+                binTypeId = currentBin.BinTypeId,
+                binTypeDesc = currentBin.BinTypeDesc,
+                cityAddress = currentBin.AreaDesc,
+                streetAddress = currentBin.BuildingAddress,
+                currentCapacity = currentBin.CurrentCapacity,
+                maxCapacity = currentBin.Capacity,
+                binTrashDisposalArea = currentBin.BinTrashDisposalArea,
+                buildingId = currentBin.BuildingId// buildingid is nullable 
+            };
+            return binData;
+        }
+
+        public Bin BinDataToDbBin(BinData binData)
+        {
+            Bin bin = new Bin()
+            {
+                BinId = binData.binId,
+                BinTypeId = binData.binTypeId,
+                CurrentCapacity = binData.currentCapacity,
+                BuildingId = binData.buildingId
+            };
+            return bin;
         }
     }
 }
@@ -312,7 +317,7 @@ namespace BL
 //                    //oldBin.CurrentCapacity = updatedBin.CurrentCapacity;
 //                    oldBin = updatedBin;
 //                    //you can update more properties here.... (in the same way)
-//                    UpdateBinLogForEachBinAction(updatedBin, dt);
+//                    AddNewBinLog(updatedBin, dt);
 //                }
 //                this.db.SaveChanges();
 //            }
@@ -322,7 +327,7 @@ namespace BL
 //            }
 //        }
 
-//        private void UpdateBinLogForEachBinAction(Bin updatedBin, DateTime dt)
+//        private void AddNewBinLog(Bin updatedBin, DateTime dt)
 //        {
 //            // TODO: Think about moving the code below into a class ... :S
 //            BinLog bLog = new BinLog();
